fix(): Break the import cycle 3
in the telegram namespace we have a cycle model.ts <-> helpers.ts. Put some helper closed to the actions. And move command helper to the command.ts module
