+++ src/telegram/commands.ts
@@ -1,4 +1,7 @@
 import type { ValueOf } from "../common/types.js";
+import type { BotMessageModel } from "./model.js";
+import type { TgMessage } from "./api/types.js";
+import { telegramBotName } from "../env.js";
 
 export const BotCommand = {
   Start: "/start",
@@ -8,3 +11,26 @@ export const BotCommand = {
 } as const;
 
 export type BotCommandType = ValueOf<typeof BotCommand>;
+
+export const isCommandMessage = (
+  model: BotMessageModel,
+  msg: TgMessage,
+  command: BotCommandType,
+): boolean => {
+  if (!msg?.text) {
+    return false;
+  }
+
+  if (msg.text === String(command)) {
+    return true;
+  }
+
+  if (!telegramBotName) {
+    return false;
+  }
+
+  return (
+    model.isGroup &&
+    msg.text.toLowerCase() === `${command}@${telegramBotName.toLowerCase()}`
+  );
+};
